ares/llm-config-generator/pyproject.toml
[project]
name = "llm-config-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm[proxy]==1.80.15",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "datamodel-code-generator[http,ruff]>=0.53.0",
    "ruff>=0.14.13",
]

[tool.ruff.lint]
select = ["I"]
